+++ lib/hui/request.ex
@@ -33,75 +33,9 @@ defmodule Hui.Request do
   @error_einval %Hui.Error{reason: :einval} # invalid argument exception
   @error_nxdomain %Hui.Error{reason: :nxdomain} # invalid / non existing host or domain
 
-  @doc """
-  Issues a search query to a specific Solr endpoint.
-
-  The query parameters can be a keyword list or a list of Hui query structs (`t:query_struct_list/0`).
-
-  ## Example - parameters
-
-  ```
-    url = "http://..."
-
-    # Parameters can be supplied as a list of keywords, which are unbound and sent to Solr directly
-    Hui.Request.search(url, q: "glen cova", facet: "true", "facet.field": ["type", "year"])
-
-    # Parameters can be a list of query structs
-    Hui.Request.search(url, [%Hui.Q{q: "glen cova"}, %Hui.F{field: ["type", "year"]}])
-
-    # DisMax query, multiple structs usage
-    x = %Hui.D{q: "edinburgh", qf: "description^2.3 title", mm: "2<-25% 9<-3"}
-    y = %Hui.Q{rows: 10, fq: ["cat:electronics"]}
-    z = %Hui.F{field: ["popularity"]} # faceting
-    Hui.Request.search(url, [x, y, z])
-  ```
-
-  The use of structs is more idiomatic and succinct. It is bound to qualified Solr fields.
-
-  The function returns a tuple or a `HTTPoison.Response` directly when `bang = true`.
-  This could be used to implement "bangified" functions such as `search!` as per Elixir convention.
-
-  ```
-    Hui.Request.search(url, [x, y, z]) 
-    # => {:ok, %HTTPoison.Response{..}}
-
-    bang = true
-    Hui.Request.search(url, bang, [x, y, z]) 
-    # => %HTTPoison.Response{..} or raise Hui.Error
-  ```
-
-  ## Example - URL endpoints
-  The URL can be specified as `t:Hui.URL.t/0`.
-
-  ```
-    url = %Hul.URL{url: "http://..."}
-    Hui.Request.search(url, q: "loch", rows: 5)
-    # -> http://.../select?q=loch&rows=5
-  ```
-
-  A key for application-configured endpoint may also be used.
-    
-  ```
-    url = :suggester
-    Hui.Request.search(url, suggest: true, "suggest.dictionary": "mySuggester", "suggest.q": "el")
-    # the above sends http://..configured_url../suggest?suggest=true&suggest.dictionary=mySuggester&suggest.q=el
-  ```
-
-  See `Hui.URL.configured_url/1` for more details.
-
-  `t:Hui.URL.t/0` struct also enables HTTP headers and HTTPoison options to be specified
-  in keyword lists. HTTPoison options provide further controls for a request, e.g. `timeout`, `recv_timeout`,
-  `max_redirect`, `params` etc.
-
-  ```
-    # setting up a header and a 10s receiving connection timeout
-    url = %Hui.URL{url: "..", headers: [{"accept", "application/json"}], options: [recv_timeout: 10000]}
-    Hui.Request.search(url, q: "solr rocks")
-  ```
-
-  See `HTTPoison.request/5` for more details on HTTPoison options.
-  """
+  @doc false
   @spec search(solr_url, boolean, solr_params) :: {:ok, HTTPoison.Response.t} | {:error, Hui.Error.t} | HTTPoison.Response.t
+  @deprecated "Please Hui.Query.get/2."
   def search(url, bang \\ false, query)
   def search(%Hui.URL{} = url, bang, query), do: _search(url, bang, query)